requirements.txt
ccxt==4.1.13
pandas==2.1.1
numpy==1.24.3
ta-lib==0.4.24
python-binance==1.0.19
python-telegram-bot==20.6
schedule==1.2.0
matplotlib==3.8.0
seaborn==0.13.0
tweepy==4.14.0
praw==7.7.1
textblob==0.17.1
requests==2.31.0
PyYAML==6.0.1
joblib==1.3.2
pytest==7.4.2
python-dotenv==1.0.0
tqdm==4.66.1
backtrader==1.9.78.123

# Web Interface Dependencies
fastapi==0.104.0
uvicorn==0.23.2
websockets==11.0.3
aiofiles==23.2.1
python-multipart==0.0.6
email-validator==2.0.0.post2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.25.0
jinja2==3.1.2
itsdangerous==2.1.2
python-socketio==5.10.0
aiohttp==3.8.6
asyncio==3.4.3

# Development Tools
black==23.9.1
isort==5.12.0
flake8==6.1.0
mypy==1.6.1
pytest-asyncio==0.21.1
pytest-cov==4.1.0
faker==19.10.0 

# Security Dependencies
cryptography==41.0.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pyjwt==2.8.0